requirements.txt
appnope==0.1.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
camel==0.1
certifi==2015.11.20.1
cycler==0.9.0
decorator==4.0.4
funcsigs==0.4
gnureadline==6.3.3
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
matplotlib==1.5.0
mock==1.3.0
networkx==1.10
neurolab==0.3.5
numpy==1.10.1
pandas==0.17.1
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
plotly==1.9.3
ptyprocess==0.5
PyBrain==0.3.3
pylab==0.1.3
pyparsing==2.0.6
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
pyzmq==15.1.0
requests==2.8.1
scikit-image==0.11.3
scikit-learn==0.17
scipy==0.16.1
seaborn==0.6.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sympy==0.7.6.1
tornado==4.3
traitlets==4.0.0
Wand==0.4.1
wheel==0.24.0
